+++ NEWS.md
@@ -1,5 +1,7 @@
 # baguette (development version)
 
+* Enable case weights for `bagger()`. 
+
 # baguette 0.2.0
 
 * Model definitions were moved to parsnip (#47)